# backend/api/gen/go/admin/goods_analytics_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             (unknown)
// source: admin/goods_analytics.proto

package admin

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
	common "shop/api/gen/go/common"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationGoodsAnalyticsServiceGetGoodsAnalyticsPie = "/admin.GoodsAnalyticsService/GetGoodsAnalyticsPie"
const OperationGoodsAnalyticsServiceGetGoodsAnalyticsSummary = "/admin.GoodsAnalyticsService/GetGoodsAnalyticsSummary"
const OperationGoodsAnalyticsServiceGetGoodsAnalyticsTrend = "/admin.GoodsAnalyticsService/GetGoodsAnalyticsTrend"

type GoodsAnalyticsServiceHTTPServer interface {
	// GetGoodsAnalyticsPie 查询商品分类分布
	GetGoodsAnalyticsPie(context.Context, *common.AnalyticsTimeRequest) (*common.AnalyticsPieResponse, error)
	// GetGoodsAnalyticsSummary 查询商品摘要指标
	GetGoodsAnalyticsSummary(context.Context, *common.AnalyticsTimeRequest) (*GoodsAnalyticsSummaryResponse, error)
	// GetGoodsAnalyticsTrend 查询商品趋势
	GetGoodsAnalyticsTrend(context.Context, *common.AnalyticsTimeRequest) (*common.AnalyticsTrendResponse, error)
}

func RegisterGoodsAnalyticsServiceHTTPServer(s *http.Server, srv GoodsAnalyticsServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/api/admin/analytics/goods/summary", _GoodsAnalyticsService_GetGoodsAnalyticsSummary0_HTTP_Handler(srv))
	r.GET("/api/admin/analytics/goods/trend", _GoodsAnalyticsService_GetGoodsAnalyticsTrend0_HTTP_Handler(srv))
	r.GET("/api/admin/analytics/goods/pie", _GoodsAnalyticsService_GetGoodsAnalyticsPie0_HTTP_Handler(srv))
}

func _GoodsAnalyticsService_GetGoodsAnalyticsSummary0_HTTP_Handler(srv GoodsAnalyticsServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in common.AnalyticsTimeRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationGoodsAnalyticsServiceGetGoodsAnalyticsSummary)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetGoodsAnalyticsSummary(ctx, req.(*common.AnalyticsTimeRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GoodsAnalyticsSummaryResponse)
		return ctx.Result(200, reply)
	}
}

func _GoodsAnalyticsService_GetGoodsAnalyticsTrend0_HTTP_Handler(srv GoodsAnalyticsServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in common.AnalyticsTimeRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationGoodsAnalyticsServiceGetGoodsAnalyticsTrend)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetGoodsAnalyticsTrend(ctx, req.(*common.AnalyticsTimeRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*common.AnalyticsTrendResponse)
		return ctx.Result(200, reply)
	}
}

func _GoodsAnalyticsService_GetGoodsAnalyticsPie0_HTTP_Handler(srv GoodsAnalyticsServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in common.AnalyticsTimeRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationGoodsAnalyticsServiceGetGoodsAnalyticsPie)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetGoodsAnalyticsPie(ctx, req.(*common.AnalyticsTimeRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*common.AnalyticsPieResponse)
		return ctx.Result(200, reply)
	}
}

type GoodsAnalyticsServiceHTTPClient interface {
	// GetGoodsAnalyticsPie 查询商品分类分布
	GetGoodsAnalyticsPie(ctx context.Context, req *common.AnalyticsTimeRequest, opts ...http.CallOption) (rsp *common.AnalyticsPieResponse, err error)
	// GetGoodsAnalyticsSummary 查询商品摘要指标
	GetGoodsAnalyticsSummary(ctx context.Context, req *common.AnalyticsTimeRequest, opts ...http.CallOption) (rsp *GoodsAnalyticsSummaryResponse, err error)
	// GetGoodsAnalyticsTrend 查询商品趋势
	GetGoodsAnalyticsTrend(ctx context.Context, req *common.AnalyticsTimeRequest, opts ...http.CallOption) (rsp *common.AnalyticsTrendResponse, err error)
}

type GoodsAnalyticsServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewGoodsAnalyticsServiceHTTPClient(client *http.Client) GoodsAnalyticsServiceHTTPClient {
	return &GoodsAnalyticsServiceHTTPClientImpl{client}
}

// GetGoodsAnalyticsPie 查询商品分类分布
func (c *GoodsAnalyticsServiceHTTPClientImpl) GetGoodsAnalyticsPie(ctx context.Context, in *common.AnalyticsTimeRequest, opts ...http.CallOption) (*common.AnalyticsPieResponse, error) {
	var out common.AnalyticsPieResponse
	pattern := "/api/admin/analytics/goods/pie"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationGoodsAnalyticsServiceGetGoodsAnalyticsPie))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetGoodsAnalyticsSummary 查询商品摘要指标
func (c *GoodsAnalyticsServiceHTTPClientImpl) GetGoodsAnalyticsSummary(ctx context.Context, in *common.AnalyticsTimeRequest, opts ...http.CallOption) (*GoodsAnalyticsSummaryResponse, error) {
	var out GoodsAnalyticsSummaryResponse
	pattern := "/api/admin/analytics/goods/summary"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationGoodsAnalyticsServiceGetGoodsAnalyticsSummary))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetGoodsAnalyticsTrend 查询商品趋势
func (c *GoodsAnalyticsServiceHTTPClientImpl) GetGoodsAnalyticsTrend(ctx context.Context, in *common.AnalyticsTimeRequest, opts ...http.CallOption) (*common.AnalyticsTrendResponse, error) {
	var out common.AnalyticsTrendResponse
	pattern := "/api/admin/analytics/goods/trend"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationGoodsAnalyticsServiceGetGoodsAnalyticsTrend))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}