# backend/internal/cmd/server/assets/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /api/admin/analytics/goods/pie:
        get:
            tags:
                - GoodsAnalyticsService
            description: 查询商品分类分布
            operationId: GoodsAnalyticsService_GetGoodsAnalyticsPie
            parameters:
                - name: timeType
                  in: query
                  schema:
                    enum:
                        - WEEK
                        - MONTH
                        - YEAR
                    type: string
                    format: enum
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AnalyticsPieResponse'
    /api/admin/analytics/goods/summary:
        get:
            tags:
                - GoodsAnalyticsService
            description: 查询商品摘要指标
            operationId: GoodsAnalyticsService_GetGoodsAnalyticsSummary
            parameters:
                - name: timeType
                  in: query
                  schema:
                    enum:
                        - WEEK
                        - MONTH
                        - YEAR
                    type: string
                    format: enum
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GoodsAnalyticsSummaryResponse'
    /api/admin/analytics/goods/trend:
        get:
            tags:
                - GoodsAnalyticsService
            description: 查询商品趋势
            operationId: GoodsAnalyticsService_GetGoodsAnalyticsTrend
            parameters:
                - name: timeType
                  in: query
                  schema:
                    enum:
                        - WEEK
                        - MONTH
                        - YEAR
                    type: string
                    format: enum
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AnalyticsTrendResponse'
...
    /api/app/shop/hot/goods:
        get:
            tags:
                - ShopHotService
            description: 查询热门推荐商品
            operationId: ShopHotService_PageShopHotGoods
            parameters:
                - name: hotItemId
                  in: query
                  description: 选项id
                  schema:
                    type: string
                - name: pageNum
                  in: query
                  description: 当前页码
                  schema:
                    type: string
                - name: pageSize
                  in: query
                  description: 每一页的行数
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PageShopHotGoodsResponse'
...
components:
    schemas:
        AnalyticsPieItem:
            type: object
            properties:
                name:
                    type: string
                    description: 名称
                value:
                    type: string
                    description: 值
        AnalyticsPieResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/AnalyticsPieItem'
                    description: 饼图数据
...
        AnalyticsTrendResponse:
            type: object
            properties:
                axis:
                    type: array
                    items:
                        type: string
                    description: 坐标轴数据
                series:
                    type: array
                    items:
                        $ref: '#/components/schemas/AnalyticsTrendSeries'
                    description: 系列列表
                yAxisNames:
                    type: array
                    items:
                        type: string
                    description: Y轴名称列表
        AnalyticsTrendSeries:
            type: object
            properties:
                name:
                    type: string
                    description: 系列名称
                type:
                    enum:
                        - BAR
                        - LINE
                    type: string
                    description: 图表类型
                    format: enum
                data:
                    type: array
                    items:
                        type: string
                    description: 数据数组
                yAxisIndex:
                    type: integer
                    description: Y轴索引
                    format: int32
...
        GoodsAnalyticsSummaryResponse:
            type: object
            properties:
                newGoodsCount:
                    type: string
                putOnGoodsRate:
                    type: string
                activeGoodsCount:
                    type: string
                activeGoodsRate:
                    type: string
                saleCount:
                    type: string
                saleGrowthRate:
                    type: string
...
        GoodsInfo:
            type: object
            properties:
                id:
                    type: string
                    description: 商品ID
                categoryId:
                    type: string
                    description: 分类ID
                name:
                    type: string
                    description: 名称
                desc:
                    type: string
                    description: 描述
                picture:
                    type: string
                    description: 商品图片
                price:
                    type: string
                    description: 当前价格(分)
                discountPrice:
                    type: string
                    description: 折扣价格（分）
                initSaleNum:
                    type: string
                    description: 初始销量
                realSaleNum:
                    type: string
                    description: 真实销售数量
                inventory:
                    type: string
                    description: 库存数量
                status:
                    enum:
                        - UNKNOWN_GS
                        - PUT_ON
                        - PULL_OFF
                    type: string
                    description: 状态
                    format: enum
                createdAt:
                    type: string
                    description: 创建时间
                updatedAt:
                    type: string
                    description: 更新时间
                categoryName:
                    type: string
                    description: 分类名称
...
        PageShopHotGoodsResponse:
            type: object
            properties:
                list:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoodsInfo'
                    description: 分页数据
                total:
                    type: integer
                    description: 总数
                    format: int32
...